requirements.txt
annotated-types==0.6.0
anyio==4.3.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
doc2pdf==0.2.0
docx2pdf==0.1.8
exceptiongroup==1.2.0
fastapi==0.109.2
google-api-core==2.17.1
google-api-python-client==2.121.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic-core==2.16.2
PyJWT==2.8.0
pyparsing==3.1.2
pypdf2==3.0.1
pytesseract==0.3.10
python-multipart==0.0.9
pywin32==306
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sniffio==1.3.0
starlette==0.36.3
tqdm==4.66.2
typing-extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.27.1
